requirements.txt
# Core framework dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# API and HTTP libraries
requests>=2.31.0
aiohttp>=3.9.0

# Google APIs
google-generativeai>=0.8.0
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# Email libraries
sendgrid>=6.11.0

# Google Sheets integration
gspread>=6.0.0
oauth2client>=4.1.3

# Data processing and utilities
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.2.0

# Type hints and validation
pydantic>=2.5.0
typing-extensions>=4.8.0

# Date/time handling
python-dateutil>=2.8.2

# UUID generation is built into Python

# JSON handling and validation
jsonschema>=4.20.0

# Development and testing (optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.8.0

# Documentation generation (optional)
mkdocs>=1.5.0
mkdocs-material>=9.5.0

# Security and encryption (for sensitive data)
cryptography>=41.0.0

# CLI utilities
click>=8.1.0
rich>=13.7.0

# Web Dashboard
Flask>=3.0.0
Flask-CORS>=6.0.0

# Database support (optional)
sqlalchemy>=2.0.0

# File handling and CSV processing
openpyxl>=3.1.0

# Error handling and retries
tenacity>=8.2.0

# Rate limiting
ratelimit>=2.2.1
